feat: download multiple files
