+++ nsdf_dark_matter_cli/src/tests/fixtures/filelist.txt
@@ -1,0 +1,10 @@
+07180808_1558_F0001
+07180808_1558_F0002
+07180808_1558_F0003
+07180808_1558_F0004
+07180808_1558_F0005
+07180808_1558_F0006
+07180808_1558_F0007
+07180808_1558_F0008
+07180808_1558_F0009
+07180808_1558_F0010
